producer/requirements.txt
kafka-python
structlog
Minio